--- brainforge/__init__.py
@@ -1,2 +1,0 @@
-from ._costs import cost as costs
-from ._activations import activation as activations
